AGA-1 : Allow setting the class for horizontal display.
Before, this iterated through all the forms to set that class.
Instead, simply use the filter 'gform_pre_form_settings_save.'
If the setting 'aga_horizontal_display' is '1',
Add a CSS class of 'gform_inline.'
@todo: only output this class in the markup.
There is no reason to see it in the form.
It's unexpected for the user to see an extra class.