AGA-1 : Change $plugin variables to $add_on.
The instantiation happens in the add-on class, Adapter_Add_On.
This is in keeping with the Gravity Forms documentation for add-ons.
Because the instantiation of classes takes place in this add-on,
Store the $add_on property in the classes, not the $plugin.
Also, add a PHPUnit test for Email_Setting::init().