AGA-1 : Admin notice if Gravity Forms isn't installed.
Instead of registering the add-on framework,
Display an admin error, indicating the minimum version needed.
Define the callback gravity_not_installed() to display this.
And store the property $_min_gravityforms_version in the class.
This is also in class Adapter_Add_On,
But this class does not have access to it.