AGA-1 : Add styling to ensure horizontal display of form.
The new versions of Gravity Forms have 2 issues with this.
When there is a name with one field, like only a first name.
Gravity Forms makes it full-width.
This makes inline display of the form impossible.
So change the width to auto.
Also, override padding in Gravity Forms browsers.css file.